auto-caller/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
passlib[bcrypt]
requests
apscheduler
retrying
retrying
